--- python/graphstorm/wholegraph.py
@@ -1,85 +1,0 @@
-"""This module provides functionality for working with the WholeGraph"""
-import os
-import json
-from typing import Optional
-
-import torch as th
-from dataclasses import dataclass
-from .utils import get_rank, get_world_size
-
-try:
-    import pylibwholegraph.torch as wgth
-except ImportError:
-    wgth = None
-
-def init_wholegraph():
-    """ Initialize Wholegraph"""
-    if wgth is None:
-        raise ImportError("WholeGraph is not installed")
-    from dgl.distributed import role
-    import pylibwholegraph.binding.wholememory_binding as wmb
-
-    @dataclass
-    class Options: # pylint: disable=missing-class-docstring
-        pass
-    Options.launch_agent = 'pytorch'
-    Options.launch_env_name_world_rank = 'RANK'
-    Options.launch_env_name_world_size = 'WORLD_SIZE'
-    Options.launch_env_name_local_rank = 'LOCAL_RANK'
-    Options.launch_env_name_local_size = 'LOCAL_WORLD_SIZE'
-    Options.launch_env_name_master_addr = 'MASTER_ADDR'
-    Options.launch_env_name_master_port = 'MASTER_PORT'
-    Options.local_rank = get_rank() % role.get_num_trainers()
-    Options.local_size = role.get_num_trainers()
-
-    wgth.distributed_launch(Options, lambda: None)
-    wmb.init(0)
-    wgth.comm.set_world_info(get_rank(), get_world_size(), Options.local_rank,
-                            Options.local_size)
-
-def load_wg_feat(part_config_path, num_parts, type_name, name):
-    """Load features from wholegraph memory
-
-    Parameters
-    ----------
-    part_config_path : str
-        The path of the partition configuration file.
-    num_parts : int
-        The number of partitions of the dataset
-    type_name: str
-        The type of node or edge for which to fetch features or labels for.
-    name: str
-        The name of the features to load
-    """
-    if wgth is None:
-        raise ImportError("WholeGraph is not installed")
-    global_comm = wgth.comm.get_global_communicator()
-    feature_comm = global_comm
-    embedding_wholememory_type = 'distributed'
-    embedding_wholememory_location = 'cpu'
-    cache_policy = wgth.create_builtin_cache_policy(
-        "none", # cache type
-        embedding_wholememory_type,
-        embedding_wholememory_location,
-        "readonly", # access type
-        0.0, # cache ratio
-    )
-    metadata_file = os.path.join(os.path.dirname(part_config_path),
-                                'wholegraph/metadata.json')
-    with open(metadata_file, encoding="utf8") as f:
-        wg_metadata = json.load(f)
-    data_shape = wg_metadata[type_name + '/' + name]['shape']
-    feat_wm_embedding = wgth.create_embedding(
-        feature_comm,
-        embedding_wholememory_type,
-        embedding_wholememory_location,
-        getattr(th, wg_metadata[type_name + '/' + name]['dtype'].split('.')[1]),
-        [data_shape[0],1] if len(data_shape) == 1 else data_shape,
-        optimizer=None,
-        cache_policy=cache_policy,
-    )
-    feat_path = os.path.join(os.path.dirname(part_config_path), 'wholegraph', \
-                                            type_name + '~' + name)
-    feat_wm_embedding.get_embedding_tensor().from_file_prefix(feat_path,
-                                                                part_count=num_parts)
-    return feat_wm_embedding
