+++ python/graphstorm/wholegraph/__init__.py
@@ -1,0 +1,19 @@
+"""
+    Copyright 2023 Contributors
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+    Initial to import WholeGraph-related function and classes
+"""
+from .wholegraph import (init_wholegraph, convert_feat_to_wholegraph, load_wg_feat)
+from .utils import is_wholegraph_embedding
